Clean up existing workflow admin
for mercycorps/TolaActivity#2822

- collect all Admin models in admin.py
- remove duplicate Admin models
